Fixing "invalid share" issue on 32-bit builds.

+++ xptMiner/global.h
@@ -195,8 +195,8 @@ void xptMiner_submitShare(minerMetiscoinBlock_t* block);
 
 // stats
 extern volatile uint64 totalCollisionCount;
-extern volatile uint64 totalShareCount;
-extern volatile uint64 invalidShareCount;
+extern volatile uint32 totalShareCount;
+extern volatile uint32 invalidShareCount;
 extern volatile uint32 monitorCurrentBlockHeight;
 
 #endif
